refactor(frontend/settings): move DynamicOptions into separate component
